Set-up linker correctly.
The linker was emitting the binary starting at 0x8000 as per usual linux
conventions. However, this meant that the ELF file it produced started with a
32KB NOP-sled. Added on top of that, the graphics card loads the kernel image at
0x8000, meaning all addresses in the image were off by 32KB (they all pointed
into the NOP-sled) causing all loads to give back zeroes.

This is now fixed, by using `objcopy` after link-time to get rid of the
NOP-sled, so the linker can write in the correct relocations, and the bootloader
can apply the offset itself.